requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
telethon==1.35.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
python-dotenv>=1.0.1
cryptography>=42.0.7
apscheduler>=3.10.4
python-multipart>=0.0.9
aiofiles>=23.2.1
Pillow>=10.0.0
pytz>=2024.1
